+++ packages/vite/src/node/http.ts
@@ -1,4 +1,4 @@
-import fs, { promises as fsp } from 'fs'
+import fs from 'fs'
 import path from 'path'
 import { Server as HttpServer } from 'http'
 import { ServerOptions as HttpsServerOptions } from 'https'
@@ -106,8 +106,7 @@ export async function resolveHttpServer(
 }
 
 export async function resolveHttpsConfig(
-  https?: boolean | HttpsServerOptions,
-  cacheDir?: string
+  https?: boolean | HttpsServerOptions
 ): Promise<HttpsServerOptions | undefined> {
   if (!https) return undefined
 
@@ -121,7 +120,7 @@ export async function resolveHttpsConfig(
     pfx: readFileIfExists(pfx)
   })
   if (!httpsOption.key || !httpsOption.cert) {
-    httpsOption.cert = httpsOption.key = await getCertificate(cacheDir)
+    httpsOption.cert = httpsOption.key = getCertificate()
   }
   return httpsOption
 }
@@ -137,104 +136,11 @@ function readFileIfExists(value?: string | Buffer | any[]) {
   return value
 }
 
-/**
- * https://github.com/webpack/webpack-dev-server/blob/master/lib/utils/createCertificate.js
- *
- * Copyright JS Foundation and other contributors
- * This source code is licensed under the MIT license found in the
- * LICENSE file at
- * https://github.com/webpack/webpack-dev-server/blob/master/LICENSE
- */
-async function createCertificate() {
-  const { generate } = await import('selfsigned')
-  const pems = generate(null, {
-    algorithm: 'sha256',
-    days: 30,
-    keySize: 2048,
-    extensions: [
-      // {
-      //   name: 'basicConstraints',
-      //   cA: true,
-      // },
-      {
-        name: 'keyUsage',
-        keyCertSign: true,
-        digitalSignature: true,
-        nonRepudiation: true,
-        keyEncipherment: true,
-        dataEncipherment: true
-      },
-      {
-        name: 'extKeyUsage',
-        serverAuth: true,
-        clientAuth: true,
-        codeSigning: true,
-        timeStamping: true
-      },
-      {
-        name: 'subjectAltName',
-        altNames: [
-          {
-            // type 2 is DNS
-            type: 2,
-            value: 'localhost'
-          },
-          {
-            type: 2,
-            value: 'localhost.localdomain'
-          },
-          {
-            type: 2,
-            value: 'lvh.me'
-          },
-          {
-            type: 2,
-            value: '*.lvh.me'
-          },
-          {
-            type: 2,
-            value: '[::1]'
-          },
-          {
-            // type 7 is IP
-            type: 7,
-            ip: '127.0.0.1'
-          },
-          {
-            type: 7,
-            ip: 'fe80::1'
-          }
-        ]
-      }
-    ]
-  })
-  return pems.private + pems.cert
-}
-
-async function getCertificate(cacheDir?: string) {
-  if (!cacheDir) return await createCertificate()
-
-  const cachePath = path.join(cacheDir, '_cert.pem')
-
-  try {
-    const [stat, content] = await Promise.all([
-      fsp.stat(cachePath),
-      fsp.readFile(cachePath, 'utf8')
-    ])
-
-    if (Date.now() - stat.ctime.valueOf() > 30 * 24 * 60 * 60 * 1000) {
-      throw new Error('cache is outdated.')
-    }
-
-    return content
-  } catch {
-    const content = await createCertificate()
-    fsp
-      .mkdir(cacheDir, { recursive: true })
-      .then(() => fsp.writeFile(cachePath, content))
-      .catch(() => {})
-    return content
-  }
+function getCertificate() {
+  return fs.readFileSync(
+    path.join(__dirname, '..', '..', '..', 'dev-cert.pem'),
+    'utf-8'
+  )
 }
 
 export async function httpServerStart(
